backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
google-auth
PyGithub
drf-yasg
gunicorn
setuptools
psycopg2
ccxt
pandas
ta-lib